requirements.txt
fastapi
uvicorn[standard]
line-bot-sdk
openai
grpcio
pillow
requests
git+https://github.com/ozgur/python-firebase